Use and plot gradients of test function derivatives; simplify variational loss function

+++ model/function_definitions.py
@@ -43,6 +43,6 @@ EXAMPLES = {
 
 # Dictionary of weight functions
 WEIGHT_FUNCTIONS = {
-    "uniform": lambda x: torch.tensor(1.0),
+    "uniform": lambda x: torch.tensor(1.0, dtype=torch.float64),
     "exponential": lambda x: torch.tensor(2.0 ** (-1 * np.sum(x))),
 }
